+++ src/test/java/dto/ScoreResultDtoTest.java
@@ -2,8 +2,8 @@ package dto;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import domain.Dealer;
-import domain.Player;
+import domain.participant.Dealer;
+import domain.participant.Player;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Suit;
